jmap_contact.c: move JSContact/vCard conversion to jscontact.c
This moves the JSContact and vCard conversion code from jmap_contact.c
to jscontact.c. This mostly is a byte-to-byte copy of the original code.
It does not change conversion, other than improving error handling.

To accomplish this, some new code had to be introduced:

- the `blobId` property now is handled by callback functions
- the `id`, `cyrusimap.org:importance` and `cyrusimap.org:href`
  properties are set in `jmap_contact.c`, not `jscontact.c`
- the UID property of a ContactCard/set update is passed in by the
  caller and stays quirk-compatible with the original code: it does not
  set the VALUE parameter for a UID that is not a URI
- blob errors are reported in the conversion context

We also drop some code that already was dead before the refactor:

- the `carddav_db` handle in the conversion context was never read
- the IGNORE_VCARD_VERSION flag had no use without SET_VCARD_CONVPROPS
  and was never passed without it
- the `record_is_dirty` boolean was a left-over from the pre-jscontact
  implementation and ignored in the JSContact code, the functions that
  that set it now return void
